relocate pytest fixtures
